Add some styling for the page. Extend the base e're setting up in common
